Point RSS link at Zola-generated RSS feed
I don't understand the intent behind that Feedly link, so this commit
might very well be completely wrong. However, this is the behavior of
the website that I observe, as currently published:
1. Clicking the RSS link brings me to a Feedly page that shows your
   content.
2. Pasting that page's URL into my feed reader subscribes me to the
   Feedly blog, not your website.

I don't know how I would use this link to subscribe to your feed, but
since the newsletter thing below the articles points to the Zola feed,
my assumption is that maybe something changed at some point, and this
Feedly link is an oversight.