Fix flashing of grid when loading page
Before this change, the main content would push the
header to the side on bigger screens.
This flickering is quite annoying and causes a redraw
with a minor performance impact. More improtantly, though
it causes an irritating flash of content on slower connections.
Fixing this by setting the size of the child elements.

The mobile view also got improved: the header now
uses less vertical space. I refactored the CSS so that
I only talk about smaller screens in the media queries
instead of mixing small and big screens.