fixing mobile styles (#66)
Cleaning up mobile styles and switching to `flex-direction: column;` on
the mobile breakpoint to put the header above the page body. Removing
`height` on the header was necessary to make the mobile breakpoint look
good, and it didn't seem to affect anything on the desktop layout. I
also removed `top` because I don't think it's doing anything now that
`position: sticky;` was removed, but I could be wrong about that.